rename: `CellViewerCanvasWrapper` -> `ViewerCanvasWrapper`
