+++ src/aics-image-viewer/components/ViewerCanvasWrapper/index.tsx
