+++ src/aics-image-viewer/components/ViewerCanvasWrapper/styles.css
